using int new scene method

+++ source/app/EditorLayer/EditorApplicationLayer.h
@@ -40,7 +40,7 @@ public:
 	bool OnRequestLoadSceneEvent(const RequestLoadSceneEvent& e);
 	bool OnRequestCloseSceneEvent(const RequestCloseSceneEvent& e);
 	bool OnDestroyScene(const OnDestroySceneEvent& e);
-	PopupResult InvokePopupRequestSaveChanges(const size_t sceneIndex);
+	PopupResult InvokePopupRequestSaveChanges(const int sceneIndex);
 	bool OnChangeActiveScene(const OnChangeActiveSceneEvent& e);
 	bool OnRequestApplicationCloseEvent(const Core::RequestApplicationCloseEvent& e);
 	bool OnApplicationCloseEvent(const Core::ApplicationCloseEvent& e);
